requirements.txt
scrapy==2.5.0
line-bot-sdk==1.19.0
sqlite3==3.35.4